revert extension file renaming
