+++ lighthouse-core/gather/connections/extension.js
@@ -158,7 +158,7 @@ class ExtensionConnection extends Connection {
   }
 
   /**
-   * Used by extension-background.js to kick off the run on the current page
+   * Used by lighthouse-ext-background to kick off the run on the current page
    * @return {Promise<string>}
    */
   getCurrentTabURL() {
