+++ lighthouse-extension/app/manifest_canary.json
@@ -12,7 +12,7 @@
   "default_locale": "en",
   "background": {
     "scripts": [
-      "scripts/extension-background.js"
+      "scripts/lighthouse-ext-background.js"
     ],
     "persistent": false
   },
