+++ lighthouse-extension/app/src/lighthouse-background.js
